Error handling + basic execution
I added error handling to make sure null checks are done on all
allocated memory (gets old after a while, to be perfectly honest). I
also wrote some code to handle parser errors. I doubt it's comprehensive
of all syntax errors that a user can make, but it seems to be doing okay
for now. I think I need to make sure it catches repeat |, &, and ; as an
error though.

The shell now also has a main loop that continuously prompts the user for
a command and executes it. It currently does not handle pipelines,
though. I need to figure out how to get a shorter working directory path
with the /home/hari replaced by ~.